+++ Sources/RepoBarCore/API/GitHubModels.swift
@@ -10,16 +10,6 @@ struct CurrentUser: Decodable {
     }
 }
 
-struct OrgMembership: Decodable {
-    let organization: Organization
-    let role: String
-    let state: String?
-
-    struct Organization: Decodable {
-        let login: String
-    }
-}
-
 struct ReleaseResponse: Decodable {
     let name: String?
     let tagName: String